+++ packages/@tinacms/mdx/package.json
@@ -14,6 +14,10 @@
 			"types": "./dist/index.d.ts",
 			"browser": "./dist/index.browser.js",
 			"default": "./dist/index.js"
+		},
+		"./sanitize-url": {
+			"types": "./dist/sanitize-url.d.ts",
+			"default": "./dist/sanitize-url.js"
 		}
 	},
 	"license": "Apache-2.0",
@@ -23,7 +27,8 @@
 				"name": "src/index.ts",
 				"target": "node",
 				"bundleDeps": true
-			}
+			},
+			"src/sanitize-url.ts"
 		]
 	},
 	"scripts": {
