+++ packages/@tinacms/mdx/src/parse/remarkToPlate.ts
@@ -8,8 +8,9 @@ import type { RichTextType } from '@tinacms/schema-tools';
 import type * as Md from 'mdast';
 import type { ContainerDirective } from 'mdast-util-directive';
 import type { MdxJsxFlowElement, MdxJsxTextElement } from 'mdast-util-mdx-jsx';
-import { directiveElement, mdxJsxElement as mdxJsxElementDefault } from './mdx';
+import { sanitizeUrl } from '../sanitize-url';
 import { parseMarkMdxText } from './mark';
+import { directiveElement, mdxJsxElement as mdxJsxElementDefault } from './mdx';
 import type * as Plate from './plate';
 
 export type { Position, PositionItem } from './plate';
@@ -600,41 +601,8 @@ export class RichTextParseError extends Error {
 }
 
 // Prevent javascript scheme (eg. `javascript:alert(document.domain)`)
-export const sanitizeUrl = (url: string | undefined) => {
-  const allowedSchemes = ['http', 'https', 'mailto', 'tel', 'xref'];
-  if (!url) return '';
-
-  let parsedUrl: URL | null = null;
-
-  try {
-    parsedUrl = new URL(url);
-  } catch (error) {
-    return url;
-  }
-
-  const scheme = parsedUrl.protocol.slice(0, -1);
-  if (allowedSchemes && !allowedSchemes.includes(scheme)) {
-    console.warn(`Invalid URL scheme detected ${scheme}`);
-    return '';
-  }
-
-  /**
-   * Trailing slash is added from new URL(...) for urls with no pathname,
-   * if the passed in url had one, keep it there, else just use the origin
-   * eg:
-   *
-   * http://example.com/ -> http://example.com/
-   * http://example.com -> http://example.com
-   * http://example.com/a/b -> http://example.com/a/b
-   * http://example.com/a/b/ -> http://example.com/a/b/
-   */
-  if (parsedUrl.pathname === '/') {
-    if (url.endsWith('/')) {
-      return parsedUrl.href;
-    }
-    // Include search (query parameters) and hash if they exist
-    return `${parsedUrl.origin}${parsedUrl.search}${parsedUrl.hash}`;
-  } else {
-    return parsedUrl.href;
-  }
-};
+// Lives in its own dependency-free file (`../sanitize-url`) so it can be
+// exported as a light `@tinacms/mdx/sanitize-url` entry point without
+// pulling in the remark/mdast/micromark toolchain. Re-exported here for
+// backwards compatibility with existing imports from this module.
+export { sanitizeUrl };
