+++ packages/tinacms/src/rich-text/index.tsx
@@ -2,7 +2,7 @@
 
 */
 
-import { sanitizeUrl } from '@tinacms/mdx';
+import { sanitizeUrl } from '@tinacms/mdx/sanitize-url';
 import React from 'react';
 
 type BaseComponents = {
